refactor: move the deterministic hash into its own module
The hash is not a property of memoization -- it digests components, vars and
rendered data under a self-delimiting encoding, and auto-memoization is just
its only caller today. Moved to reflex_base/utils/deterministic_hash.py with
the tests alongside it.

The private _deterministic_hash/_update_deterministic_hash pair becomes one
public variadic deterministic_hash(*values), which is what the two call sites
wanted: component_hash now reads as the render plus the artifacts that identify
a memo body, and _update_component_artifacts_hash becomes _component_artifacts,
a generator that yields them instead of threading a hasher and buffer through.
One shared buffer still covers the whole digest.

Nothing else under utils imports components at runtime, so the two isinstance
checks that need Var and BaseComponent import them inside _resolve_hash_encoder
-- once per type, so it never shows up in a profile -- and the module now
imports standalone without pulling in the component system.

Digests are unchanged: memo tags across three benchmark pages are byte-identical
to before the move, and an A/B of the encoder before and after runs at parity.
