+++ tests/units/components/test_memo.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import dataclasses
 import inspect
 import re
 from collections.abc import Callable
@@ -13,8 +12,6 @@ from unittest.mock import patch
 import pytest
 from reflex_base.components.component import Component
 from reflex_base.components.memo import (
-    _HASH_MAX_CACHE_ENTRIES,
-    _HASH_MAX_CACHED_DATACLASS,
     _SPECS,
     DEFAULT_MEMO_WRAPPER,
     EMPTY_VAR_COMPONENT,
@@ -25,19 +22,12 @@ from reflex_base.components.memo import (
     MemoParam,
     MemoParamKind,
     _analyze_params,
-    _deterministic_hash,
-    _hash_dataclass_encodings,
-    _hash_dataclass_layouts,
-    _hash_encoders,
-    _hash_str_encodings,
     _LazyBody,
     _MemoCallBinding,
     _strip_optional,
-    clear_hash_caches,
     component_hash,
     memo_tag,
 )
-from reflex_base.constants import Hooks
 from reflex_base.event import EventChain, EventHandler, no_args_event_spec
 from reflex_base.registry import RegistrationContext
 from reflex_base.style import Style
@@ -2054,126 +2044,6 @@ def test_self_referencing_var_memo():
     assert "recursive_count" in str(invoked)
 
 
-@pytest.fixture
-def clean_hash_caches():
-    """Isolate a test from the module-level memo-naming caches.
-
-    Tests that fill these caches would otherwise leave their probe values in
-    place for the rest of the session, and tests run in random order, so a test
-    that reads cache state has to start from a known one.
-
-    Yields:
-        None, with the caches empty on entry and on exit.
-    """
-    clear_hash_caches()
-    yield
-    clear_hash_caches()
-
-
-def test_deterministic_hash_is_stable():
-    """The same value must hash identically across calls and dict orderings."""
-    value = {"b": [1, "x"], "a": {"k": None}}
-    reordered = {"a": {"k": None}, "b": [1, "x"]}
-
-    assert _deterministic_hash(value) == _deterministic_hash(value)
-    assert _deterministic_hash(value) == _deterministic_hash(reordered)
-
-
-@pytest.mark.parametrize(
-    ("left", "right"),
-    [
-        # Type tags must keep values of different types apart.
-        ("1", 1),
-        (1, True),
-        (0, False),
-        (None, "None"),
-        ({"a": "b"}, [["a", "b"]]),
-        # Length prefixes must keep concatenations apart.
-        (["ab", "c"], ["a", "bc"]),
-        ([[], []], [[[]]]),
-        ({"a": "", "b": ""}, {"ab": ""}),
-        # Nested containers must not flatten into their contents.
-        ([1, [2]], [1, 2]),
-        # str-keyed enums encode as enums, not as their string value.
-        (Hooks.HookPosition.PRE_TRIGGER, Hooks.HookPosition.PRE_TRIGGER.value),
-        # Dataclasses of the same shape but different types stay distinct.
-        (ImportVar(tag="a"), ImportVar(tag="a", alias="a")),
-    ],
-)
-def test_deterministic_hash_distinguishes(left: Any, right: Any):
-    """Distinct values must not collide under the type-tagged encoding."""
-    assert _deterministic_hash(left) != _deterministic_hash(right)
-
-
-def test_deterministic_hash_treats_lists_and_tuples_alike():
-    """Sequences share one type tag, so a list and tuple of equal items match."""
-    assert _deterministic_hash([1, "a"]) == _deterministic_hash((1, "a"))
-
-
-def test_deterministic_hash_import_var_cache_is_by_value():
-    """Equal ``ImportVar`` instances hash the same; unequal ones do not.
-
-    ``ImportVar`` encodings are cached by value, so a stale or over-eager cache
-    entry would show up as two unequal imports hashing alike.
-    """
-    a = ImportVar(tag="useState", is_default=False, install=True)
-    b = ImportVar(tag="useState", is_default=False, install=True)
-    c = ImportVar(tag="useState", is_default=True, install=True)
-
-    assert _deterministic_hash(a) == _deterministic_hash(b)
-    assert _deterministic_hash(a) != _deterministic_hash(c)
-    assert _deterministic_hash({"react": (a, c)}) != _deterministic_hash({
-        "react": (c, a)
-    })
-
-
-def test_deterministic_hash_long_strings():
-    """Strings past the encoding cache's size limit still hash correctly."""
-    long_a = "a" * 10_000
-    long_b = "a" * 9_999 + "b"
-
-    assert _deterministic_hash(long_a) == _deterministic_hash("a" * 10_000)
-    assert _deterministic_hash(long_a) != _deterministic_hash(long_b)
-
-
-def test_deterministic_hash_beyond_string_cache_capacity(clean_hash_caches: None):
-    """Strings that arrive after the encoding cache fills still hash correctly."""
-    values = [f"cache_capacity_probe_{i}" for i in range(_HASH_MAX_CACHE_ENTRIES + 500)]
-    digests = [_deterministic_hash(value) for value in values]
-
-    assert len(set(digests)) == len(values)
-    assert [_deterministic_hash(value) for value in values] == digests
-
-
-def test_deterministic_hash_flushes_large_payloads():
-    """A payload larger than the buffer flush size hashes deterministically."""
-    payload = {f"key_{i}": "v" * 200 for i in range(2000)}
-
-    assert _deterministic_hash(payload) == _deterministic_hash(dict(payload))
-    mutated = {**payload, "key_0": "w" * 200}
-    assert _deterministic_hash(payload) != _deterministic_hash(mutated)
-
-
-def test_deterministic_hash_components_and_vars():
-    """Components and Vars hash by rendered content, not by identity."""
-    assert _deterministic_hash(Bare.create(contents="a")) == _deterministic_hash(
-        Bare.create(contents="a")
-    )
-    assert _deterministic_hash(Bare.create(contents="a")) != _deterministic_hash(
-        Bare.create(contents="b")
-    )
-    assert _deterministic_hash(Var("a")) == _deterministic_hash(Var("a"))
-    assert _deterministic_hash(Var("a")) != _deterministic_hash(Var("b"))
-    # A Var and the bare string it renders to must not collide.
-    assert _deterministic_hash(Var("a")) != _deterministic_hash("a")
-
-
-def test_deterministic_hash_rejects_unsupported_types():
-    """Values with no encoding raise rather than hashing to a shared digest."""
-    with pytest.raises(TypeError):
-        _deterministic_hash(object())
-
-
 class _AppWrapProbe(Component):
     """A component whose only per-instance artifact is an app-wrap component."""
 
@@ -2217,175 +2087,6 @@ def test_component_hash_covers_dataclass_inheriting_app_wrap_components():
     assert component_hash(a, recursive=False) != component_hash(b, recursive=False)
     assert component_hash(a, recursive=True) != component_hash(b, recursive=True)
     assert memo_tag(a) != memo_tag(b)
-
-
-def test_deterministic_hash_encodes_dataclass_components_as_components():
-    """A component that also inherits a dataclass encodes its render."""
-    assert _deterministic_hash(rx.text("a")) != _deterministic_hash(rx.text("b"))
-    # A Var is a frozen dataclass too, and must keep encoding as a Var.
-    assert _deterministic_hash(Var("a")) != _deterministic_hash(Var("b"))
-
-
-@dataclasses.dataclass(frozen=True)
-class _KeyedProbe:
-    """A frozen dataclass whose declared fields are all safe to key on."""
-
-    name: str
-    flag: bool = False
-    alias: str | None = None
-
-
-@dataclasses.dataclass(frozen=True)
-class _DefaultsProbe:
-    """A frozen dataclass every field of which has a default."""
-
-    name: str = "default"
-
-
-class _PlainProbe:
-    """A plain class -- no encoding, and ``_DefaultsProbe``'s metaclass."""
-
-
-@dataclasses.dataclass(frozen=True)
-class _NumericProbe:
-    """A frozen dataclass with a field that ``==`` can conflate."""
-
-    value: int | bool
-
-
-@dataclasses.dataclass(frozen=True)
-class _ContainerProbe:
-    """A frozen dataclass holding something that can still change."""
-
-    items: list[int]
-
-
-@dataclasses.dataclass
-class _MutableProbe:
-    """An unfrozen dataclass, and so an unhashable cache key."""
-
-    value: str
-
-
-def test_deterministic_hash_caches_any_keyable_frozen_dataclass(
-    clean_hash_caches: None,
-):
-    """Frozen dataclasses of str/bool/None fields are cached, not just imports."""
-    probe = _KeyedProbe(name="probe")
-
-    digest = _deterministic_hash(probe)
-    assert _hash_dataclass_encodings.get(probe) is not None
-    # An equal instance reuses the entry and lands on the same digest; unequal
-    # ones must not.
-    assert _deterministic_hash(_KeyedProbe(name="probe")) == digest
-    assert _deterministic_hash(_KeyedProbe(name="probe", flag=True)) != digest
-    assert _deterministic_hash(_KeyedProbe(name="other")) != digest
-
-
-def test_deterministic_hash_does_not_cache_numeric_frozen_dataclasses(
-    clean_hash_caches: None,
-):
-    """A field that can hold a number is not keyable by value.
-
-    ``True == 1`` and the two hash alike, so a cache keyed on the instance would
-    hand ``_NumericProbe(1)`` the encoding of ``_NumericProbe(True)``.
-    """
-    boolean, numeric = _NumericProbe(value=True), _NumericProbe(value=1)
-
-    assert boolean == numeric
-    assert _deterministic_hash(boolean) != _deterministic_hash(numeric)
-    assert not _hash_dataclass_encodings
-
-
-def test_deterministic_hash_tracks_dataclasses_that_can_still_change():
-    """Dataclasses whose contents can change must be re-encoded every time."""
-    mutable = _MutableProbe(value="before")
-    digest = _deterministic_hash(mutable)
-    mutable.value = "after"
-    assert _deterministic_hash(mutable) != digest
-
-    # Frozen, but a field holds a mutable container.
-    container = _ContainerProbe(items=[1])
-    digest = _deterministic_hash(container)
-    container.items.append(2)
-    assert _deterministic_hash(container) != digest
-
-
-def test_deterministic_hash_handles_dataclasses_without_params():
-    """Classes that copy ``__dataclass_fields__`` without the decorator hash.
-
-    ``MutableProxy`` synthesizes its wrapper classes exactly this way:
-    ``dataclasses.is_dataclass`` is true, but ``__dataclass_params__`` never
-    comes along, so there is no ``frozen`` flag to read.
-    """
-    synthesized = type(
-        "_SynthesizedProbe",
-        (),
-        {
-            "__dataclass_fields__": _KeyedProbe.__dataclass_fields__,
-            "name": "probe",
-            "flag": False,
-            "alias": None,
-        },
-    )
-
-    assert _deterministic_hash(synthesized()) == _deterministic_hash(
-        _KeyedProbe(name="probe")
-    )
-
-
-def test_deterministic_hash_skips_oversized_dataclass_encodings(
-    clean_hash_caches: None,
-):
-    """Outsized encodings are not retained, keeping the cache's memory bounded."""
-    oversized = _KeyedProbe(name="x" * _HASH_MAX_CACHED_DATACLASS)
-
-    digest = _deterministic_hash(oversized)
-    assert not _hash_dataclass_encodings
-    assert _deterministic_hash(_KeyedProbe(name="x" * _HASH_MAX_CACHED_DATACLASS)) == (
-        digest
-    )
-
-
-def test_deterministic_hash_beyond_dataclass_cache_capacity(clean_hash_caches: None):
-    """Frozen dataclasses arriving after the cache fills still hash correctly."""
-    values = [
-        _KeyedProbe(name=f"capacity_probe_{i}")
-        for i in range(_HASH_MAX_CACHE_ENTRIES + 100)
-    ]
-    digests = [_deterministic_hash(value) for value in values]
-
-    assert len(set(digests)) == len(values)
-    assert [_deterministic_hash(value) for value in values] == digests
-
-
-def test_deterministic_hash_encoder_table_keeps_types_apart(clean_hash_caches: None):
-    """Memoizing an encoder per type must not route other types through it."""
-    # A str-keyed enum resolves to the enum encoder; plain strings must keep
-    # their own fast path, and the two must not collide.
-    position = Hooks.HookPosition.PRE_TRIGGER
-    assert _deterministic_hash(position) != _deterministic_hash(position.value)
-    assert _deterministic_hash(position.value) == _deterministic_hash(
-        str(position.value)
-    )
-    # Dataclass types are encoded from their defaults. A class's own type is
-    # its metaclass, which it shares with unrelated classes, so caching an
-    # encoder under it would send those down the dataclass path too.
-    assert type(_DefaultsProbe) is type(_PlainProbe)
-    assert _deterministic_hash(_DefaultsProbe) == _deterministic_hash(
-        _DefaultsProbe(name="default")
-    )
-    with pytest.raises(TypeError, match="Cannot hash value"):
-        _deterministic_hash(_PlainProbe)
-
-
-def test_deterministic_hash_unsupported_type_is_not_memoized(clean_hash_caches: None):
-    """A type with no encoder must raise every time, not just the first."""
-    with pytest.raises(TypeError, match="Cannot hash value"):
-        _deterministic_hash(object())
-    with pytest.raises(TypeError, match="Cannot hash value"):
-        _deterministic_hash(object())
-    assert not _hash_encoders
 
 
 class _CustomCodeProbe(Component):
@@ -2528,37 +2229,3 @@ def test_memo_tag_separates_identically_rendering_classes():
 
     assert alpha.render() == beta.render()
     assert memo_tag(alpha) != memo_tag(beta)
-
-
-def test_clear_hash_caches_drops_every_cache(clean_hash_caches: None):
-    """The compile-scoped encoding caches must all be released together.
-
-    Nothing asks for these values after a compile, and a dataclass type defined
-    inside a function body is a fresh class object each time -- so a cache left
-    behind would pin one per compile for the life of the process.
-    """
-    ephemeral = dataclasses.make_dataclass("Ephemeral", [("v", str)])
-    before = _deterministic_hash({
-        "prop": ephemeral(v="x"),
-        "imports": (ImportVar(tag="useCacheProbe"),),
-    })
-
-    assert _hash_dataclass_layouts
-    assert _hash_str_encodings
-    assert _hash_dataclass_encodings
-    assert _hash_encoders
-
-    clear_hash_caches()
-
-    assert not _hash_dataclass_layouts
-    assert not _hash_str_encodings
-    assert not _hash_dataclass_encodings
-    assert not _hash_encoders
-    # Hashing rebuilds them from scratch and must land on the same digest.
-    assert (
-        _deterministic_hash({
-            "prop": ephemeral(v="x"),
-            "imports": (ImportVar(tag="useCacheProbe"),),
-        })
-        == before
-    )
